feat(181): remove placeholder
aip.dev/181 discusses stability levels of the APIs themselves.

There have been discusses on and off about adopting this AIP, but
there has not yet been consensus on whether it is valuable to adopt:
it effectively provides a restriction on versioning of the APIs that
is not materially related to the clients.

In addition, some AEP maintainers are considering a separate proposal
that introduces date-based versioning schemes for APIs to handle
introducition and removal of fields. Since that may supercede
this proposal in the future, that is another reason to not keep a placeholder 
value for an aip that may ultimately have separate guidance.