fix(pattern,oas): add _id suffix to resource ids (#321)
The current OAS guidance requires using the structure `{singular}_id`, 
but the examples currently do not include this suffix.

This also adjusts the convention in the `pattern` field of the `x-aep-resource` annotation as well. 
This ensures consistency in pathing structure documentation.

See https://github.com/aep-dev/aeps/discussions/320#discussioncomment-14123990 for 
a discussion.

---------
Co-authored-by: Richard Frankel <richard@frankel.tv>